requirements.txt
python-dotenv
pandas
numpy
scikit-learn
sentence-transformers
openai
pypdf
tqdm
pytest
